refactor: Rename filenames of layout components
